change scale to f32

+++ src/run.rs
@@ -8,7 +8,7 @@ use gridsim::{GetNeighbors, Sim, SquareGrid, TakeMoveNeighbors};
 use std::fs::File;
 use Renderer;
 
-const SCALED_DEFAULT: u32 = 10;
+const SCALED_DEFAULT: f32 = 10.0;
 
 /// Runs a grid with default window setup. Draws true as white and false as black.
 #[cfg_attr(feature = "flame_it", flame)]
@@ -28,7 +28,7 @@ where
 
 // Runs a grid with each cell in the window is scale pixels^2. Draws true as white and false as black.
 #[cfg_attr(feature = "flame_it", flame)]
-pub fn scaled_bool<'a, S: 'a>(grid: SquareGrid<'a, S>, scale: u32)
+pub fn scaled_bool<'a, S: 'a>(grid: SquareGrid<'a, S>, scale: f32)
 where
     S: Sim<'a, Cell = bool>,
     S::Cell: Sync + Send,
@@ -61,7 +61,7 @@ where
 
 /// Runs a grid with each cell in the window is scale pixels^2 and a coloration function.
 #[cfg_attr(feature = "flame_it", flame)]
-pub fn scaled<'a, S: 'a, F>(grid: SquareGrid<'a, S>, coloration: F, scale: u32)
+pub fn scaled<'a, S: 'a, F>(grid: SquareGrid<'a, S>, coloration: F, scale: f32)
 where
     S: Sim<'a>,
     F: Fn(&S::Cell) -> [f32; 4] + Sync,
@@ -103,7 +103,7 @@ pub fn scaled_filter<'a, S: 'a, Color, Filter>(
     mut grid: SquareGrid<'a, S>,
     coloration: Color,
     filter: Filter,
-    scale: u32,
+    scale: f32,
 ) where
     S: Sim<'a>,
     Color: Fn(&S::Cell) -> [f32; 4] + Sync,
@@ -117,7 +117,10 @@ pub fn scaled_filter<'a, S: 'a, Color, Filter>(
     SquareGrid<'a, S>: GetNeighbors<'a, usize, S::Neighbors>,
 {
     let mut events_loop = glutin::EventsLoop::new();
-    let window = glutin::WindowBuilder::new().with_dimensions(scale * grid.get_width() as u32, scale * grid.get_height() as u32);
+    let window = glutin::WindowBuilder::new().with_dimensions(
+        (scale * grid.get_width() as f32) as u32,
+        (scale * grid.get_height() as f32) as u32,
+    );
     let context = glutin::ContextBuilder::new().with_vsync(true);
     let display = glium::Display::new(window, context, &events_loop).unwrap();
     let renderer = Renderer::new(&display);
